This is a spike of simplifying the logic used by the ListEvaluator. It pushes the responsibility of driving the evaluation process to the actual procedures and macros of the system, while hiding the internals of calling conventions and activation structures behind an execution context.
It currently does not handle the lambda's responsiblity of returning the
last evaluation result. FIXME's are indicated throughout the source.

The ListEvaluator now only cares about whether the head item evaluated to a
Macro or not. If it does, it is given control over the current execution
environment (to provide the custom list processing), otherwise list evaluation
occurs as normal and a procedure is expected to exist in the head position to
apply at the end of evaluation. This eliminates the painful switch logic that
was implemented before and uses polymorphic message dispatch to handle the
different evaluation cases.

Other notes:
* Many scheme objects return null from the new
SchemeObject evaluate(ExecutionContext) method. This method should be void and all
returns pushed through the ExecutionContext.returnValue methods
* Similarly many implementations simply want to return a value without looking at
the execution context. A mechanism for simply returning a value from a Java method
based on the provided scheme environment would be nice.
* Calling apply() on types that are not "applyable" throws a RuntimeException. This
should be a proper exception.
* The ExecutionContext interface grew organically from a blank slate. Each method
was added as use-cases arose. As such much of its functionality is simply a
thin wrapper around the SchemeStack and ActivationFrame. I think it helped to come
up with an interface that didn't deal directly with "stacks" and "frames". I can
envisage coming up with an implementation of the ExecutionContext that mimics
our original native Java stack implementation by simply evaluating the next object
directly, rather than manipulating our own stack structures.

This is still a WIP, with many rough edges but I think it sets the direction for
the rest of the work we want to do.